requirements.txt
dnspython==2.1.0
newsapi-python==0.2.6
pymongo==3.11.4